backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
httpx
python-dotenv 
bcrypt
python-jose[cryptography]
python-dotenv